Extract to a side-effect module
