+++ ui-next/src/index.ts
@@ -6,6 +6,10 @@
  * that enterprise packages can use to extend the application.
  */
 
+// Pins the Monaco CDN version for consumers of this package. Must stay a
+// side-effect import so it runs before any editor component mounts.
+import "./monacoLoader";
+
 // =============================================================================
 // Plugin System - Primary export for enterprise extensions
 // =============================================================================
